requirements.txt
ConfigArgParse==1.0
dhash==1.3
Flask==1.1.1
PyExifTool==0.1.1
Send2Trash==1.5.0
Wand==0.5.8
PyYAML==5.3